moved nfts in collection widget
