+++ src/pages/CollectionPage/CollectionPage.scss
@@ -6,29 +6,3 @@
         padding: 0;
     }
 }
-
-.nft-list {
-    display: grid;
-    grid-template-columns: repeat(auto-fill, minmax(16rem, 1fr));
-    justify-items: center;
-    row-gap: 1.5rem;
-    margin: 1.5rem;
-}
-
-.status {
-    @extend %flex-align-center; 
-    width: 100%;
-    height: 19rem;
-    font-size: 1rem;
-    font-weight: 700;
-    color: var(--c-white);
-}
-
-.loading {
-    @extend %flex-align-center; 
-    width:100%;
-    height: 8rem;
-    font-size: 1rem;
-    font-weight: 700;
-    color: var(--c-white);
-}
